+++ src/SoftTissue.DataContract/ExperimentalAnalysis/AnalyzeAndExtendResults/AnalyzeAndExtendResultsResponse.cs
@@ -1,9 +1,0 @@
-﻿using SoftTissue.DataContract.OperationBase;
-
-namespace SoftTissue.DataContract.ExperimentalAnalysis.AnalyzeAndExtendResults
-{
-    /// <summary>
-    /// It represents the response content for AnalyzeAndPredictResults operations.
-    /// </summary>
-    public sealed class AnalyzeAndExtendResultsResponse : OperationResponseBase<AnalyzeAndExtendResultsResponseData> { }
-}