+++ src/SoftTissue.DataContract/ExperimentalAnalysis/AnalyzeAndExtrapolateResults/AnalyzeAndExtrapolateResultsResponse.cs
@@ -1,0 +1,9 @@
+﻿using SoftTissue.DataContract.OperationBase;
+
+namespace SoftTissue.DataContract.ExperimentalAnalysis.AnalyzeAndExtrapolateResults
+{
+    /// <summary>
+    /// It represents the response content for AnalyzeAndPredictResults operations.
+    /// </summary>
+    public sealed class AnalyzeAndExtrapolateResultsResponse : OperationResponseBase<AnalyzeAndExtrapolateResultsResponseData> { }
+}